update to minecrunch_launcher
moved GetResources and MinecraftCheck to Utility class as it is more
appropriate in that class
